requirements.txt
tweepy>=4.14
python-dotenv
matplotlib
pandas
requests
plotly
kaleido
